+++ backend/src/app.js
@@ -3,6 +3,7 @@ import cors from "cors";
 import { router as authRouter } from "./routes/authRouter.js";
 import { authenticateToken } from "./middleware/authMiddleware.js";
 import { router as taskRouter } from "./routes/taskRouter.js";
+import { router as dashboardRouter} from "./routes/dashboardRouter.js";
 
 //* Getting the App from Express and adding cors and json middleware
 const app = express();
@@ -12,5 +13,6 @@ app.use(express.json());
 //* App routes
 app.use("/auth", authRouter);
 app.use("/tasks", authenticateToken, taskRouter);
+app.use("/",authenticateToken,dashboardRouter);
 
 export default app;
