+++ backend/src/routes/dashboardRouter.js
@@ -1,0 +1,11 @@
+import { Router } from "express";
+import { getDashboard } from "../controllers/dashboardController.js";
+
+const router = Router();
+
+router.get("/",getDashboard);
+
+
+export {
+    router
+};
